split off json to own files
