+++ merlict_c89/mliSurface_json.h
@@ -1,0 +1,15 @@
+/* Copyright 2018-2020 Sebastian Achim Mueller */
+#ifndef MERLICT_C89_MLISURFACE_JSON_H_
+#define MERLICT_C89_MLISURFACE_JSON_H_
+
+#include "mliSurface.h"
+#include "mliJson.h"
+
+int mliSurface_malloc_from_json_str(
+        struct mliSurface *surface,
+        const char *json_str);
+int mliSurface_malloc_from_json_token(
+        struct mliSurface *surface,
+        const struct mliJson *json,
+        const uint64_t token);
+#endif
